requirements.txt
alembic==1.9.1
aniso8601==9.0.1
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
coverage==7.0.3
dparse==0.6.2
exceptiongroup==1.1.0
flake8==6.0.0
flasgger==0.9.5
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
iniconfig==1.1.1
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
mistune==2.0.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.0
pytest-cov==4.0.0
pytest-sugar==0.9.6
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
safety==2.3.5
six==1.16.0
SQLAlchemy==1.4.46
termcolor==2.2.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
